aider: Add conditional check to include `args` and `kwargs` in the `request_data` dictionary as a field called `request`, only if they are not of type `UploadFile`, otherwise add the `serialize_request_data` object.
# Aider chat conversation:

USER: add the args and kwargs to the request_data as a field called request, only if they are not of type `UploadFile`, if they are of type `UploadFile` then add the `serialize_request_data` object

ASSISTANT: To implement the requested changes in the `airtable.py` file, we will modify the `log_endpoint` decorator function. We will add a conditional check to determine if `args` and `kwargs` contain any `UploadFile` instances. If they do not contain `UploadFile` instances, we will add them to the `request_data` dictionary under the key `request`. If there are `UploadFile` instances, we will add the `serialize_request_data` object instead.

Here is the *SEARCH/REPLACE block* for the required change:

airtable.py
```python
<<<<<<< SEARCH
        request_data = {
            "action": func.__name__
        }
=======
        request_data = {
            "action": func.__name__,
            "request": serialize_request_data({"args": args, "kwargs": kwargs}) if any(isinstance(arg, UploadFile) for arg in args) or any(isinstance(v, UploadFile) for v in kwargs.values()) else {"args": args, "kwargs": kwargs}
        }
>>>>>>> REPLACE
```

This change will add a new field `request` to the `request_data` dictionary, which will either contain the serialized request data if there are `UploadFile` instances or the raw `args` and `kwargs` if there are no `UploadFile` instances.